# openapi_v3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "angie",
    "version": "1.0.0",
    "license": {
      "name": "SEE LICENSE IN LICENSE"
    },
    "description": "Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3).\n\nAngie's "
  },
  "paths": {
...
    "/integration/{id}/channel/{channel}/deploy": {
      "post": {
        "summary": "/integration/{id}/channel/{channel}/deploy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "integration"
        ]
      }
    },
    "/integration/{id}/channel/{channel}/undeploy": {
      "post": {
        "summary": "/integration/{id}/channel/{channel}/undeploy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "integration"
        ]
      }
    },
    "/integration/{id}/channel/{channel}/status": {
      "get": {
        "summary": "/integration/{id}/channel/{channel}/status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "integration"
        ]
      },
      "post": {
        "summary": "/integration/{id}/channel/{channel}/status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "integration"
        ]
      }
    },
    "/integration/{id}/channel/{channel}/log": {
      "get": {
        "summary": "/integration/{id}/channel/{channel}/log",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "integration"
        ]
      }
    },
    "/integration_channel/to_camel": {
      "post": {
        "summary": "/integration_channel/to_camel",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string",
                      "example": "<routes  xmlns=\"http://camel.apache.org/schema/spring\"><route> <from uri=\"mllp://0.0.0.0:8888\"/> <multicast> <to uri=\"direct:e9d0cc1a-ce57-4125-9857-d5a89a3eedad\"/> </multicast></route><route> <from uri=\"direct:e9d0cc1a-ce57-4125-9857-d5a89a3eedad\"/> <setBody>  <groovy>import ca.uhn.hl7v2.model.v25.group.ORM_O01_ORDER;import ca.uhn.hl7v2.model.v25.message.ORM_O01;\r\ndef msg = request.getBody(ca.uhn.hl7v2.model.Message.class);\r\ndef terser = new ca.uhn.hl7v2.util.Terser(msg);\r\ndef tests = [:];ORM_O01 orm_o01 = (ORM_O01) msg;\r\ndef counter = 0;\r\nfor (ORM_O01_ORDER orderObservation : orm_o01.getORDERAll()) {\r\n    def orc = orderObservation.getORC();\r\n    def obr = orderObservation.getORDER_DETAIL().getOBR();\r\n    tests[counter++] = [\r\n        \"test_code\":obr.getObr4_UniversalServiceIdentifier().getCe1_Identifier().getValue(),\r\n        \"test_name\":obr.getObr4_UniversalServiceIdentifier().getCe2_Text().getValue(),\r\n        \"sample_key\": orc.getOrc4_PlacerGroupNumber().getName(),\r\n        \"sample_name\": null,\r\n        \"container\": null,\r\n        \"result\": null,\r\n        \"result_date\": null\r\n    ]\r\n};\r\ndef json = [\r\n    order_number_source: terser.get(\"/ORDER/ORC-2-1\"),\r\n    order_number: terser.get(\"/ORDER/ORC-2-1\"),order_comment: null,\r\n    order_status: null,\r\n    order_date: terser.get(\"/MSH-7-1\"),\r\n    order_source: terser.get(\"/MSH-3-1\"),\r\n    nif: terser.get(\"/PATIENT/PID-20-1\"),\r\n    nhc: null, firstname: terser.get(\"/PATIENT/PID-5-2\"),\r\n    lastname: terser.get(\"/PATIENT/PID-5-1\"),\r\n    lastname2: null,\r\n    birthdate: terser.get(\"/PATIENT/PID-7-1\"),\r\n    address: terser.get(\"/PATIENT/PID-11-1\"),\r\n    sex: terser.get(\"/PATIENT/PID-8-1\"),\r\n    phone: terser.get(\"/PATIENT/PID-13-1\"),\r\n    physician: terser.get(\"/ORDER/ORC-12-1-1\"),\r\n    physician_name: terser.get(\"/PATIENT/PATIENT_VISIT/PV1-9-1-1\") + terser.get(\"/PATIENT/PATIENT_VISIT/PV1-9-2-1\"),\r\n    physician_esp_name: terser.get(\"/PATIENT/PATIENT_VISIT/PV1-20-2-1\"),\r\n    domain: terser.get(\"/PATIENT/PATIENT_VISIT/PV1-2\"),\r\n    type: terser.get(\"/PATIENT/PATIENT_VISIT/PV1-2\"),\r\n    area: null,\r\n    area_name: null,\r\n    priority: null,\r\n    bed: null,\r\n    collpoint: null,\r\n    inscomp: terser.get(\"/PATIENT/PATIENT_VISIT/PV1-20-1-1\"),\r\n    collpoint_name: null,\r\n    tests: tests\r\n];  return new com.fasterxml.jackson.databind.ObjectMapper().writeValueAsString(json);</groovy>    </setBody> <multicast> <to uri=\"direct:f67afb3a-a817-4ae6-988d-1f802b98cf96\"/> </multicast> </route><route> <from uri=\"direct:cb4b97ba-2228-45ce-b019-07fd39504d34\"/> <to uri=\"log:debug\"/> </route><route> <from uri=\"direct:f67afb3a-a817-4ae6-988d-1f802b98cf96\"/> <setBody>  <>import ca.uhn.hl7v2.model.v25.group.ORM_O01_ORDER\r\nimport groovy.json.JsonSlurper\r\nimport groovy.json.JsonParserType\r\n\r\n//Login to SL\r\ndef msg = request.getBody()\r\ndef post = new URL(\"http://10.10.1.34/system/login/\").openConnection()\r\ndef message = '{\"usr\":\"suser\",\"pwd\":\"smartlis\"}'\r\n\r\npost.setRequestMethod(\"POST\")\r\npost.setDoOutput(true)\r\npost.setRequestProperty(\"Content-Type\", \"application/json\")\r\npost.getOutputStream().write(message.getBytes(\"UTF-8\"))\r\ndef postRC = post.getResponseCode()\r\n\r\ndef parser = new JsonSlurper()\r\ndef parser1 = new JsonSlurper()\r\nif (postRC.equals(200)) { //If Login Succesfull\r\n\r\n    def session = post.getInputStream().getText()\r\n    def json = parser.parseText(session)\r\n    msg = parser1.parseText(msg)\r\n    def exec = [\"job_info\": [\"name\"  : \"ORDER_CREATE: \" + msg[\"order_number\"],\"param1\": msg[\"order_number\"]],\"work_sess_id\": json[\"data\"][\"new_sess_key\"],\"space_code\": \"\",\"exec\": [:]]\r\n    def commandCounter = 1\r\n    def newOrder = [\"cmd\": \"new\", \"id\": msg[\"order_number\"]]\r\n    exec[\"exec\"][commandCounter++] =  newOrder\r\n    def newOrderCommand = [\"cmd\": \"modify_order_data\", \"field\": \"order_source\", \"data\": json[\"order_source\"]]\r\n    exec[\"exec\"][commandCounter++] =  newOrderCommand\r\n    def addOrderDemographicCommand1 = [\"cmd\": \"modify_order_data\", \"field\": \"physician_name\", \"data\": json[\"physician_name\"]]\r\n    exec[\"exec\"][commandCounter++] =  addOrderDemographicCommand1\r\n    def addOrderDemographicCommand2 = [\"cmd\": \"modify_order_data\", \"field\": \"physician\", \"data\": json[\"physician\"]]\r\n    exec[\"exec\"][commandCounter++] =  addOrderDemographicCommand2\r\n    def addOrderDemographicCommand3 = [\"cmd\": \"modify_order_data\", \"field\": \"pat_name\", \"data\": json[\"lastname\"]]\r\n    exec[\"exec\"][commandCounter++] = addOrderDemographicCommand3 \r\n    def addOrderDemographicCommand4 = [\"cmd\": \"modify_order_data\", \"field\": \"pat_name\", \"data\": json[\"lastname\"]]\r\n    exec[\"exec\"][commandCounter++] =  addOrderDemographicCommand4\r\n    def addOrderDemographicCommand11 = [\"cmd\": \"modify_order_data\", \"field\": \"source\", \"data\": msg[\"order_source\"]]\r\n    exec[\"exec\"][commandCounter++] =  addOrderDemographicCommand11\r\n    def col1lp1oint = [\"cmd\"  : \"modify_order_data\",\"field\": \"collpoint\",\"data\" : \"5006\"]\r\n    exec[\"exec\"][commandCounter++] =  col1lp1oint\r\n    def ins1comp = [\"cmd\"  : \"modify_order_data\",\"field\": \"inscomp\",\"data\" : \"CEY-262\"]\r\n    exec[\"exec\"][commandCounter++] =  ins1comp\r\n    def coll_name = [\"cmd\"  : \"modify_order_data\",\"field\": \"collpoint_name\",\"data\" : msg[\"order_source\"]]\r\n    exec[\"exec\"][commandCounter++] =  coll_name\r\n    def source = [\"cmd\"  : \"modify_order_data\",\"field\": \"source\",\"data\" : msg[\"order_source\"]]\r\n    exec[\"exec\"][commandCounter++] =  source\r\n\r\n    if (msg[\"tests\"] != null) {\r\n        for (def genTestIndex in msg[\"tests\"]) {\r\n            def genTest =  genTestIndex.collect([] as HashSet){entry -> entry.value}\r\n             \r\n            def addGenTestCommand = [\"cmd\": \"add_test2\", \"test_code\":genTestIndex.collect({entry -> entry.value.test_code})[0] ]\r\n            \r\n            if (genTest &amp;&amp; genTest[\"comment\"] != null &amp;&amp; genTest[\"comment\"] != \"\") {\r\n            \r\n            }\r\n            exec[\"exec\"][commandCounter++] =  addGenTestCommand\r\n        }\r\n    }\r\n\r\n    def saveCommand = [\"cmd\": \"save\", \"async_mode\": \"false\"]\r\n    exec[\"exec\"][commandCounter++] =  saveCommand\r\n    return new com.fasterxml.jackson.databind.ObjectMapper().writeValueAsString(exec)\r\n}</>    </setBody> <multicast> <to uri=\"direct:cb4b97ba-2228-45ce-b019-07fd39504d34\"/> </multicast> </route></routes>"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "total": {
                      "type": "number",
                      "example": 1
                    },
                    "message": {
                      "type": "string",
                      "example": ""
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Missing helper: \"camelSafe\""
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "integration"
        ]
      }
    },
    "/channel/{id}/sendMessageToRoute": {
      "post": {
        "summary": "/channel/{id}/sendMessageToRoute",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": []
      }
    },
...
    "/section/{role}/check_allowed": {
      "post": {
        "summary": "/section/{role}/check_allowed",
        "parameters": [
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/section/{role}/path_allowed": {
      "post": {
        "summary": "/section/{role}/path_allowed",
        "parameters": [
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": []
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://localhost:3105"
    }
  ],
  "components": {
    "securitySchemes": {
      "authorization": {
        "name": "authorization",
        "in": "header",
        "type": "apiKey"
      }
    }
  }
}